2to3: lib2to3.fixes.fix_long
svn path=/ZMS/branches/zms4/; revision=4221

+++ _confmanager.py
@@ -63,7 +63,7 @@ class ConfDict:
     @classmethod
     def get(cls):
         if cls.__confdict__ is None:
-            cls.__confdict__ = {'last_modified':long(DateTime().timeTime())}
+            cls.__confdict__ = {'last_modified':int(DateTime().timeTime())}
             PRODUCT_HOME = os.path.dirname(os.path.abspath(__file__))
             for home in [PRODUCT_HOME,INSTANCE_HOME]:
               fp = os.path.join(home,'etc','zms.conf')
